// src/app/modules/master-forms/Models/Stockist-Branch-Model.ts
export class StockistBranchModel {

  BranchId: number = 0;
  CompId: number = 0;
  StockistId: number = 0;
  CompName:string='';
  Stockist: string = '';
  Branch: string = '';
  Company: string = '';
  BranchStr:string='';
  CompStr:string='';
  CompanyStr:string='';
  Addedby: string = '';
  Action: string = '';
  StockistName:string='';
  BranchName:string='';
  Stockieststr:string='';
}